[Feature #158743386] Refactor code
